Adjusted index references in Interpolate to better handle fixed and variable arguments. ProcessDetailsFactory::init now throws an exception if the supplied destination is the wrong count. Updated Choice::getExpression to ensure that the choice is an integer. Improved CollectionFeatures::shape to support shape(-1). TraversalPolicy::toStringDetail now distinguishes between fixed and variable count. Removed unused code.

+++ algebra/src/main/java/org/almostrealism/algebra/computations/Choice.java
@@ -39,7 +39,7 @@ public class Choice<T extends PackedCollection<?>> extends TraversableExpression
 	protected CollectionExpression getExpression(TraversableExpression... args) {
 		return CollectionExpression.create(getShape(), idx -> {
 			Expression choice = args[1].getValueAt(e(0)).multiply(choiceCount).floor();
-			Expression pos = choice.multiply(getShape().getSize());
+			Expression pos = choice.toInt().multiply(getShape().getSize());
 			return args[2].getValueAt(pos.add(idx));
 		});
 	}
